fix Bug of closing when moving Task in Board
sopra-fs24-group07/sopra-fs24-group-07-client/issues/74
-now update Task when opened does not work anymore :( sorry basil
