+++ src/components/ui/ColumnContainer.tsx
@@ -1,19 +1,33 @@
-import React, { useMemo, useState } from "react";
+import React, { useState } from "react";
 import PropTypes from "prop-types";
 import "../../styles/ui/ColumnContainer.scss";
 import { Button } from "./Button";
 import TaskCard from "./TaskCard";
 import { useDroppable } from "@dnd-kit/core";
-import { CSS } from "@dnd-kit/utilities";
 import CreateTask from "../popups/CreateTask";
+import InspectTask from "components/popups/InspectTask";
 
 function ColumnContainer(props) {
   //get the columns and tasks from props
   const { column, teamTasks, sessionStatus, isDragged } = props;
   //new Task that can be created
-  const [newTask, setNewTask] = useState(null);
+  const [openTask, setOpenTask] = useState(null);
   //set State of Create Task Popup
   const [isCreateTaskOpen, setCreateTaskOpen] = useState(false);
+
+  const [isInspectTaskOpen, setInspectTaskOpen] = useState(false);
+
+  //open the Inspect Task Popup
+  const openInspectTask = (task) => {
+    setOpenTask(task);
+    setInspectTaskOpen(true);
+  };
+
+  //close the Inspect Task Popup
+  const closeInspectTask = () => {
+    console.log("I WAS IN THERE");
+    setInspectTaskOpen(false);
+  };
 
   //dnd
   const { isOver, setNodeRef } = useDroppable({
@@ -52,6 +66,7 @@ function ColumnContainer(props) {
 
             return (
               <TaskCard
+                openInspectTask={openInspectTask}
                 key={task.id}
                 task={task}
                 col={column}
@@ -75,6 +90,14 @@ function ColumnContainer(props) {
           </Button>
           <CreateTask isOpen={isCreateTaskOpen} onClose={closeCreateTask} />
         </div>
+      )}
+      {openTask && (
+        <InspectTask
+          isOpen={isInspectTaskOpen}
+          onClose={closeInspectTask}
+          task={openTask}
+          inSession={false}
+        />
       )}
     </div>
   );
